refactor: 일반 로그인, 카카오 로그인 화면 하나로 통합 - 카카오와 일반로그인은 controller만 userController로 통합 - 나머지는 일반 user와 카카오 kakao로 분리 docs: application-secret.yml - kakao api 부분 수정 docs: EticaPlannerApplication 부분 추가 - @SpringBootApplication(exclude={SecurityAutoConfiguration.class}) - 스프링 시큐리티 추가할때 자동으로 나오는 화면을 제거하기 위함 - 자동설정으로인해 로그인을 입력하지 않을시 어떠한 url로도 이동할 수 없음 그래서 제거
