+++ eticaPlanner/src/main/java/com/eticaplanner/eticaPlanner/kakao/Entity/KakaoUserEntity.java
@@ -14,14 +14,15 @@ import lombok.*;
 public class KakaoUserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @Column(name = "kakao_no")
+    private Long id; // // Entity에서의 id는 실제 DB의 'kakao_no'에 매핑됨
 
     @Column(name = "kakao_id", nullable = false, unique = true)
     private Long  kakaoId;
 
-    @Column(name = "nickname", nullable = false)
+    @Column(name = "kakao_nickname", nullable = false)
     private String nickname;
 
-    @Column(name = "email", unique = true, nullable = false)
+    @Column(name = "kakao_email", unique = true, nullable = false)
     private String email;
 }
